+++ History.md
@@ -9,3 +9,8 @@
 	-- 然后TaskVent
 	-- 最后开启一个或多个TaskWorker
 	-- 结论： 开启一个worker用时5039ms，而开启2个用时1700，判别为并行处理模型
+	-- 注意：需要先开启Client，然后开启Server（需要connect Client），最后才能connect Client和Server绑定的端口
+* Comment():有这样的说法：绑定套接字至端点；连接套接字至端点。端点指的是某个广为周知网络地址
+			DEALER过去被称为XREQ，ROUTER被称为XREP
+* Add: ZMQ多线程编程: MTServer - MTClient
+	   请求-应答代理: RRServer - RRClient